Unified UX for answer correctness
